Reorginize the repo a bit from what cvs2svn imported
git-svn-id: https://swig.svn.sourceforge.net/svnroot/swig/trunk2/SWIG@9590 626c5289-ae23-0410-ae9c-e8d60b6d4f22

+++ Examples/GIFPlot/Python/full/runme.py
@@ -1,5 +1,7 @@
 # Plot a 3D function
-from gifplot import *
+# This example uses the low-level C interface.
+
+from _gifplot import *
 from math import *
 
 # Here is the function to plot
@@ -47,7 +49,7 @@ def drawsolid():
 			c2 = cscale*(z2-zmin)
 			c3 = cscale*(z3-zmin)
 			c4 = cscale*(z4-zmin)
-			c = (c1+c2+c3+c4)/4
+			c = int((c1+c2+c3+c4)/4)
 			if (c < 0) : c = 0
 			if c > 239 : c = 239
 			Plot3D_solidquad(p3,x,y,z1,x+dx,y,z2,x+dx,y+dy,z3,x,y+dy,z4,c+16)
